Inline social auth controller actions
